+++ pybit/_v5_market.py
@@ -3,6 +3,19 @@ from .market import Market
 
 
 class MarketHTTP(_V5HTTPManager):
+    def get_server_time(self) -> dict:
+        """
+        Returns:
+            Request results as dictionary.
+
+        Additional information:
+            https://bybit-exchange.github.io/docs/v5/market/time
+        """
+        return self._submit_request(
+            method="GET",
+            path=f"{self.endpoint}{Market.GET_SERVER_TIME}",
+        )
+
     def get_kline(self, **kwargs) -> dict:
         """Query the kline data. Charts are returned in groups based on the requested interval.